scraper/requirements.txt
playwright>=1.40.0
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
requests>=2.31.0
anthropic>=0.40.0
streamlit>=1.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
